Moved Index validation to json::Root
